// apps/api/src/achievements/types/achievements.types.ts
export enum AchievementsTypes {
    WELCOME = 'WELCOME',
    TASTEOFV = 'TASTEOFV',
    WINS10PLUS = 'WINS10PLUS',
    WIN10RAW = 'WIN10RAW',
    NEMESIS = 'NEMESIS',
    FIRST = 'FIRST',
    POINTS100PLUS = 'POINTS100PLUS',
    WINSINLESSTHAN1M = 'WINSINLESSTHAN1M',
    ENDLESSSTAMINA = 'ENDLESSSTAMINA'
}